% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{AchievementDefinition}
\alias{AchievementDefinition}
\title{Google Play Game Services API Objects 
The API for Google Play Game Services.}
\usage{
AchievementDefinition(achievementType = NULL, description = NULL,
  experiencePoints = NULL, formattedTotalSteps = NULL, id = NULL,
  initialState = NULL, isRevealedIconUrlDefault = NULL,
  isUnlockedIconUrlDefault = NULL, name = NULL, revealedIconUrl = NULL,
  totalSteps = NULL, unlockedIconUrl = NULL)
}
\arguments{
\item{achievementType}{The type of the achievement}

\item{description}{The description of the achievement}

\item{experiencePoints}{Experience points which will be earned when unlocking this achievement}

\item{formattedTotalSteps}{The total steps for an incremental achievement as a string}

\item{id}{The ID of the achievement}

\item{initialState}{The initial state of the achievement}

\item{isRevealedIconUrlDefault}{Indicates whether the revealed icon image being returned is a default image, or is provided by the game}

\item{isUnlockedIconUrlDefault}{Indicates whether the unlocked icon image being returned is a default image, or is game-provided}

\item{name}{The name of the achievement}

\item{revealedIconUrl}{The image URL for the revealed achievement icon}

\item{totalSteps}{The total steps for an incremental achievement}

\item{unlockedIconUrl}{The image URL for the unlocked achievement icon}
}
\value{
AchievementDefinition object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2017-03-05 19:52:02
filename: /Users/mark/dev/R/autoGoogleAPI/googlegamesv1.auto/R/games_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
AchievementDefinition Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for an achievement definition object.
}
