% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{AchievementDefinitionsListResponse}
\alias{AchievementDefinitionsListResponse}
\title{AchievementDefinitionsListResponse Object}
\usage{
AchievementDefinitionsListResponse(items = NULL, nextPageToken = NULL)
}
\arguments{
\item{items}{The achievement definitions}

\item{nextPageToken}{Token corresponding to the next page of results}
}
\value{
AchievementDefinitionsListResponse object
}
\description{
AchievementDefinitionsListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for a list of achievement definition objects.
}

