% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{AchievementIncrementResponse}
\alias{AchievementIncrementResponse}
\title{AchievementIncrementResponse Object}
\usage{
AchievementIncrementResponse(currentSteps = NULL, newlyUnlocked = NULL)
}
\arguments{
\item{currentSteps}{The current steps recorded for this incremental achievement}

\item{newlyUnlocked}{Whether the current steps for the achievement has reached the number of steps required to unlock}
}
\value{
AchievementIncrementResponse object
}
\description{
AchievementIncrementResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for an achievement increment response
}

