% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{AchievementRevealResponse}
\alias{AchievementRevealResponse}
\title{AchievementRevealResponse Object}
\usage{
AchievementRevealResponse(currentState = NULL)
}
\arguments{
\item{currentState}{The current state of the achievement for which a reveal was attempted}
}
\value{
AchievementRevealResponse object
}
\description{
AchievementRevealResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for an achievement reveal response
}

