% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_functions.R
\name{achievements.increment}
\alias{achievements.increment}
\title{Increments the steps of the achievement with the given ID for the currently authenticated player.}
\usage{
achievements.increment(achievementId, stepsToIncrement,
  consistencyToken = NULL, requestId = NULL)
}
\arguments{
\item{achievementId}{The ID of the achievement used by this method}

\item{stepsToIncrement}{The number of steps to increment}

\item{consistencyToken}{The last-seen mutation timestamp}

\item{requestId}{A randomly generated numeric ID for each request specified by the caller}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/games
\item https://www.googleapis.com/auth/plus.login
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/games, https://www.googleapis.com/auth/plus.login)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/games/services/}{Google Documentation}
}
