% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_functions.R
\name{achievements.setStepsAtLeast}
\alias{achievements.setStepsAtLeast}
\title{Sets the steps for the currently authenticated player towards unlocking an achievement. If the steps parameter is less than the current number of steps that the player already gained for the achievement, the achievement is not modified.}
\usage{
achievements.setStepsAtLeast(achievementId, steps, consistencyToken = NULL)
}
\arguments{
\item{achievementId}{The ID of the achievement used by this method}

\item{steps}{The minimum value to set the steps to}

\item{consistencyToken}{The last-seen mutation timestamp}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/games
\item https://www.googleapis.com/auth/plus.login
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/games, https://www.googleapis.com/auth/plus.login)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/games/services/}{Google Documentation}
}
