% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{AchievementSetStepsAtLeastResponse}
\alias{AchievementSetStepsAtLeastResponse}
\title{AchievementSetStepsAtLeastResponse Object}
\usage{
AchievementSetStepsAtLeastResponse(currentSteps = NULL,
  newlyUnlocked = NULL)
}
\arguments{
\item{currentSteps}{The current steps recorded for this incremental achievement}

\item{newlyUnlocked}{Whether the the current steps for the achievement has reached the number of steps required to unlock}
}
\value{
AchievementSetStepsAtLeastResponse object
}
\description{
AchievementSetStepsAtLeastResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for an achievement set steps at least response.
}
