% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{AchievementUnlockResponse}
\alias{AchievementUnlockResponse}
\title{AchievementUnlockResponse Object}
\usage{
AchievementUnlockResponse(newlyUnlocked = NULL)
}
\arguments{
\item{newlyUnlocked}{Whether this achievement was newly unlocked (that is, whether the unlock request for the achievement was the first for the player)}
}
\value{
AchievementUnlockResponse object
}
\description{
AchievementUnlockResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for an achievement unlock response
}

