% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{AchievementUpdateMultipleRequest}
\alias{AchievementUpdateMultipleRequest}
\title{AchievementUpdateMultipleRequest Object}
\usage{
AchievementUpdateMultipleRequest(updates = NULL)
}
\arguments{
\item{updates}{The individual achievement update requests}
}
\value{
AchievementUpdateMultipleRequest object
}
\description{
AchievementUpdateMultipleRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for a list of achievement update requests.
}
\seealso{
Other AchievementUpdateMultipleRequest functions: \code{\link{achievements.updateMultiple}}
}

