% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{AchievementUpdateMultipleResponse}
\alias{AchievementUpdateMultipleResponse}
\title{AchievementUpdateMultipleResponse Object}
\usage{
AchievementUpdateMultipleResponse(updatedAchievements = NULL)
}
\arguments{
\item{updatedAchievements}{The updated state of the achievements}
}
\value{
AchievementUpdateMultipleResponse object
}
\description{
AchievementUpdateMultipleResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for an achievement unlock response.
}

