% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{AchievementUpdateRequest}
\alias{AchievementUpdateRequest}
\title{AchievementUpdateRequest Object}
\usage{
AchievementUpdateRequest(achievementId = NULL, incrementPayload = NULL,
  setStepsAtLeastPayload = NULL, updateType = NULL)
}
\arguments{
\item{achievementId}{The achievement this update is being applied to}

\item{incrementPayload}{The payload if an update of type INCREMENT was requested for the achievement}

\item{setStepsAtLeastPayload}{The payload if an update of type SET_STEPS_AT_LEAST was requested for the achievement}

\item{updateType}{The type of update being applied}
}
\value{
AchievementUpdateRequest object
}
\description{
AchievementUpdateRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for a request to update an achievement.
}
