% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{AchievementUpdateResponse}
\alias{AchievementUpdateResponse}
\title{AchievementUpdateResponse Object}
\usage{
AchievementUpdateResponse(achievementId = NULL, currentState = NULL,
  currentSteps = NULL, newlyUnlocked = NULL, updateOccurred = NULL)
}
\arguments{
\item{achievementId}{The achievement this update is was applied to}

\item{currentState}{The current state of the achievement}

\item{currentSteps}{The current steps recorded for this achievement if it is incremental}

\item{newlyUnlocked}{Whether this achievement was newly unlocked (that is, whether the unlock request for the achievement was the first for the player)}

\item{updateOccurred}{Whether the requested updates actually affected the achievement}
}
\value{
AchievementUpdateResponse object
}
\description{
AchievementUpdateResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for an achievement update response.
}

