% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{AggregateStats}
\alias{AggregateStats}
\title{AggregateStats Object}
\usage{
AggregateStats(count = NULL, max = NULL, min = NULL, sum = NULL)
}
\arguments{
\item{count}{The number of messages sent between a pair of peers}

\item{max}{The maximum amount}

\item{min}{The minimum amount}

\item{sum}{The total number of bytes sent for messages between a pair of peers}
}
\value{
AggregateStats object
}
\description{
AggregateStats Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for aggregate stats.
}
