% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{AnonymousPlayer}
\alias{AnonymousPlayer}
\title{AnonymousPlayer Object}
\usage{
AnonymousPlayer(avatarImageUrl = NULL, displayName = NULL)
}
\arguments{
\item{avatarImageUrl}{The base URL for the image to display for the anonymous player}

\item{displayName}{The name to display for the anonymous player}
}
\value{
AnonymousPlayer object
}
\description{
AnonymousPlayer Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for an anonymous player
}

