% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{ApplicationCategory}
\alias{ApplicationCategory}
\title{ApplicationCategory Object}
\usage{
ApplicationCategory(primary = NULL, secondary = NULL)
}
\arguments{
\item{primary}{The primary category}

\item{secondary}{The secondary category}
}
\value{
ApplicationCategory object
}
\description{
ApplicationCategory Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for an application category object.
}
