% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_functions.R
\name{applications.get}
\alias{applications.get}
\title{Retrieves the metadata of the application with the given ID. If the requested application is not available for the specified platformType, the returned response will not include any instance data.}
\usage{
applications.get(applicationId, consistencyToken = NULL, language = NULL,
  platformType = NULL)
}
\arguments{
\item{applicationId}{The application ID from the Google Play developer console}

\item{consistencyToken}{The last-seen mutation timestamp}

\item{language}{The preferred language to use for strings returned by this method}

\item{platformType}{Restrict application details returned to the specific platform}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/games
\item https://www.googleapis.com/auth/plus.login
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/games, https://www.googleapis.com/auth/plus.login)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/games/services/}{Google Documentation}
}
