% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{ApplicationVerifyResponse}
\alias{ApplicationVerifyResponse}
\title{ApplicationVerifyResponse Object}
\usage{
ApplicationVerifyResponse(alternate_player_id = NULL, player_id = NULL)
}
\arguments{
\item{alternate_player_id}{An alternate ID that was once used for the player that was issued the auth token used in this request}

\item{player_id}{The ID of the player that was issued the auth token used in this request}
}
\value{
ApplicationVerifyResponse object
}
\description{
ApplicationVerifyResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for a third party application verification response resource.
}
