% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{Category}
\alias{Category}
\title{Category Object}
\usage{
Category(category = NULL, experiencePoints = NULL)
}
\arguments{
\item{category}{The category name}

\item{experiencePoints}{Experience points earned in this category}
}
\value{
Category object
}
\description{
Category Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for data related to individual game categories.
}
