% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{CategoryListResponse}
\alias{CategoryListResponse}
\title{CategoryListResponse Object}
\usage{
CategoryListResponse(items = NULL, nextPageToken = NULL)
}
\arguments{
\item{items}{The list of categories with usage data}

\item{nextPageToken}{Token corresponding to the next page of results}
}
\value{
CategoryListResponse object
}
\description{
CategoryListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for a list of category data objects.
}

