% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{EventBatchRecordFailure}
\alias{EventBatchRecordFailure}
\title{EventBatchRecordFailure Object}
\usage{
EventBatchRecordFailure(failureCause = NULL, range = NULL)
}
\arguments{
\item{failureCause}{The cause for the update failure}

\item{range}{The time range which was rejected; empty for a request-wide failure}
}
\value{
EventBatchRecordFailure object
}
\description{
EventBatchRecordFailure Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for a batch update failure resource.
}

