% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{EventChild}
\alias{EventChild}
\title{EventChild Object}
\usage{
EventChild(childId = NULL)
}
\arguments{
\item{childId}{The ID of the child event}
}
\value{
EventChild object
}
\description{
EventChild Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for an event child relationship resource.
}
