% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{EventDefinition}
\alias{EventDefinition}
\title{EventDefinition Object}
\usage{
EventDefinition(childEvents = NULL, description = NULL,
  displayName = NULL, id = NULL, imageUrl = NULL,
  isDefaultImageUrl = NULL, visibility = NULL)
}
\arguments{
\item{childEvents}{A list of events that are a child of this event}

\item{description}{Description of what this event represents}

\item{displayName}{The name to display for the event}

\item{id}{The ID of the event}

\item{imageUrl}{The base URL for the image that represents the event}

\item{isDefaultImageUrl}{Indicates whether the icon image being returned is a default image, or is game-provided}

\item{visibility}{The visibility of event being tracked in this definition}
}
\value{
EventDefinition object
}
\description{
EventDefinition Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for an event definition resource.
}

