% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{EventDefinitionListResponse}
\alias{EventDefinitionListResponse}
\title{EventDefinitionListResponse Object}
\usage{
EventDefinitionListResponse(items = NULL, nextPageToken = NULL)
}
\arguments{
\item{items}{The event definitions}

\item{nextPageToken}{The pagination token for the next page of results}
}
\value{
EventDefinitionListResponse object
}
\description{
EventDefinitionListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for a ListDefinitions response.
}
