% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{EventPeriodRange}
\alias{EventPeriodRange}
\title{EventPeriodRange Object}
\usage{
EventPeriodRange(periodEndMillis = NULL, periodStartMillis = NULL)
}
\arguments{
\item{periodEndMillis}{The time when this update period ends, in millis, since 1970 UTC (Unix Epoch)}

\item{periodStartMillis}{The time when this update period begins, in millis, since 1970 UTC (Unix Epoch)}
}
\value{
EventPeriodRange object
}
\description{
EventPeriodRange Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for an event period time range.
}
