% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{EventPeriodUpdate}
\alias{EventPeriodUpdate}
\title{EventPeriodUpdate Object}
\usage{
EventPeriodUpdate(timePeriod = NULL, updates = NULL)
}
\arguments{
\item{timePeriod}{The time period being covered by this update}

\item{updates}{The updates being made for this time period}
}
\value{
EventPeriodUpdate object
}
\description{
EventPeriodUpdate Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for an event period update resource.
}
