% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{EventRecordFailure}
\alias{EventRecordFailure}
\title{EventRecordFailure Object}
\usage{
EventRecordFailure(eventId = NULL, failureCause = NULL)
}
\arguments{
\item{eventId}{The ID of the event that was not updated}

\item{failureCause}{The cause for the update failure}
}
\value{
EventRecordFailure object
}
\description{
EventRecordFailure Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for an event update failure resource.
}

