% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{EventRecordRequest}
\alias{EventRecordRequest}
\title{EventRecordRequest Object}
\usage{
EventRecordRequest(currentTimeMillis = NULL, requestId = NULL,
  timePeriods = NULL)
}
\arguments{
\item{currentTimeMillis}{The current time when this update was sent, in milliseconds, since 1970 UTC (Unix Epoch)}

\item{requestId}{The request ID used to identify this attempt to record events}

\item{timePeriods}{A list of the time period updates being made in this request}
}
\value{
EventRecordRequest object
}
\description{
EventRecordRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for an event period update resource.
}
\seealso{
Other EventRecordRequest functions: \code{\link{events.record}}
}
