% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_functions.R
\name{events.listByPlayer}
\alias{events.listByPlayer}
\title{Returns a list showing the current progress on events in this application for the currently authenticated user.}
\usage{
events.listByPlayer(consistencyToken = NULL, language = NULL,
  maxResults = NULL, pageToken = NULL)
}
\arguments{
\item{consistencyToken}{The last-seen mutation timestamp}

\item{language}{The preferred language to use for strings returned by this method}

\item{maxResults}{The maximum number of events to return in the response, used for paging}

\item{pageToken}{The token returned by the previous request}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/games
\item https://www.googleapis.com/auth/plus.login
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/games, https://www.googleapis.com/auth/plus.login)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/games/services/}{Google Documentation}
}
