% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{EventUpdateRequest}
\alias{EventUpdateRequest}
\title{EventUpdateRequest Object}
\usage{
EventUpdateRequest(definitionId = NULL, updateCount = NULL)
}
\arguments{
\item{definitionId}{The ID of the event being modified in this update}

\item{updateCount}{The number of times this event occurred in this time period}
}
\value{
EventUpdateRequest object
}
\description{
EventUpdateRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for an event period update resource.
}

