% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{EventUpdateResponse}
\alias{EventUpdateResponse}
\title{EventUpdateResponse Object}
\usage{
EventUpdateResponse(batchFailures = NULL, eventFailures = NULL,
  playerEvents = NULL)
}
\arguments{
\item{batchFailures}{Any batch-wide failures which occurred applying updates}

\item{eventFailures}{Any failures updating a particular event}

\item{playerEvents}{The current status of any updated events}
}
\value{
EventUpdateResponse object
}
\description{
EventUpdateResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for an event period update resource.
}
