% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_functions.R
\docType{package}
\name{games_googleAuthR}
\alias{games_googleAuthR}
\alias{games_googleAuthR-package}
\title{Google Play Game Services API
The API for Google Play Game Services.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2016-09-03 23:13:41
filename: /Users/mark/dev/R/autoGoogleAPI/googlegamesv1.auto/R/games_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/drive.appdata
\item https://www.googleapis.com/auth/games
\item https://www.googleapis.com/auth/plus.login
}
}

