% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{GamesAchievementIncrement}
\alias{GamesAchievementIncrement}
\title{GamesAchievementIncrement Object}
\usage{
GamesAchievementIncrement(requestId = NULL, steps = NULL)
}
\arguments{
\item{requestId}{The requestId associated with an increment to an achievement}

\item{steps}{The number of steps to be incremented}
}
\value{
GamesAchievementIncrement object
}
\description{
GamesAchievementIncrement Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for the payload to request to increment an achievement.
}

