% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{ImageAsset}
\alias{ImageAsset}
\title{ImageAsset Object}
\usage{
ImageAsset(height = NULL, name = NULL, url = NULL, width = NULL)
}
\arguments{
\item{height}{The height of the asset}

\item{name}{The name of the asset}

\item{url}{The URL of the asset}

\item{width}{The width of the asset}
}
\value{
ImageAsset object
}
\description{
ImageAsset Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for an image asset object.
}

