% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{Instance}
\alias{Instance}
\title{Instance Object}
\usage{
Instance(acquisitionUri = NULL, androidInstance = NULL,
  iosInstance = NULL, name = NULL, platformType = NULL,
  realtimePlay = NULL, turnBasedPlay = NULL, webInstance = NULL)
}
\arguments{
\item{acquisitionUri}{URI which shows where a user can acquire this instance}

\item{androidInstance}{Platform dependent details for Android}

\item{iosInstance}{Platform dependent details for iOS}

\item{name}{Localized display name}

\item{platformType}{The platform type}

\item{realtimePlay}{Flag to show if this game instance supports realtime play}

\item{turnBasedPlay}{Flag to show if this game instance supports turn based play}

\item{webInstance}{Platform dependent details for Web}
}
\value{
Instance object
}
\description{
Instance Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for the Instance resource.
}

