% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{InstanceAndroidDetails}
\alias{InstanceAndroidDetails}
\title{InstanceAndroidDetails Object}
\usage{
InstanceAndroidDetails(enablePiracyCheck = NULL, packageName = NULL,
  preferred = NULL)
}
\arguments{
\item{enablePiracyCheck}{Flag indicating whether the anti-piracy check is enabled}

\item{packageName}{Android package name which maps to Google Play URL}

\item{preferred}{Indicates that this instance is the default for new installations}
}
\value{
InstanceAndroidDetails object
}
\description{
InstanceAndroidDetails Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for the Android instance details resource.
}
