% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{InstanceIosDetails}
\alias{InstanceIosDetails}
\title{InstanceIosDetails Object}
\usage{
InstanceIosDetails(bundleIdentifier = NULL, itunesAppId = NULL,
  preferredForIpad = NULL, preferredForIphone = NULL, supportIpad = NULL,
  supportIphone = NULL)
}
\arguments{
\item{bundleIdentifier}{Bundle identifier}

\item{itunesAppId}{iTunes App ID}

\item{preferredForIpad}{Indicates that this instance is the default for new installations on iPad devices}

\item{preferredForIphone}{Indicates that this instance is the default for new installations on iPhone devices}

\item{supportIpad}{Flag to indicate if this instance supports iPad}

\item{supportIphone}{Flag to indicate if this instance supports iPhone}
}
\value{
InstanceIosDetails object
}
\description{
InstanceIosDetails Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for the iOS details resource.
}
