% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{InstanceWebDetails}
\alias{InstanceWebDetails}
\title{InstanceWebDetails Object}
\usage{
InstanceWebDetails(launchUrl = NULL, preferred = NULL)
}
\arguments{
\item{launchUrl}{Launch URL for the game}

\item{preferred}{Indicates that this instance is the default for new installations}
}
\value{
InstanceWebDetails object
}
\description{
InstanceWebDetails Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for the Web details resource.
}
