% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{Leaderboard}
\alias{Leaderboard}
\title{Leaderboard Object}
\usage{
Leaderboard(iconUrl = NULL, id = NULL, isIconUrlDefault = NULL,
  name = NULL, order = NULL)
}
\arguments{
\item{iconUrl}{The icon for the leaderboard}

\item{id}{The leaderboard ID}

\item{isIconUrlDefault}{Indicates whether the icon image being returned is a default image, or is game-provided}

\item{name}{The name of the leaderboard}

\item{order}{How scores are ordered}
}
\value{
Leaderboard object
}
\description{
Leaderboard Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for the Leaderboard resource.
}
