% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{LeaderboardEntry}
\alias{LeaderboardEntry}
\title{LeaderboardEntry Object}
\usage{
LeaderboardEntry(formattedScore = NULL, formattedScoreRank = NULL,
  player = NULL, scoreRank = NULL, scoreTag = NULL, scoreValue = NULL,
  timeSpan = NULL, writeTimestampMillis = NULL)
}
\arguments{
\item{formattedScore}{The localized string for the numerical value of this score}

\item{formattedScoreRank}{The localized string for the rank of this score for this leaderboard}

\item{player}{The player who holds this score}

\item{scoreRank}{The rank of this score for this leaderboard}

\item{scoreTag}{Additional information about the score}

\item{scoreValue}{The numerical value of this score}

\item{timeSpan}{The time span of this high score}

\item{writeTimestampMillis}{The timestamp at which this score was recorded, in milliseconds since the epoch in UTC}
}
\value{
LeaderboardEntry object
}
\description{
LeaderboardEntry Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for the Leaderboard Entry resource.
}
