% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{LeaderboardListResponse}
\alias{LeaderboardListResponse}
\title{LeaderboardListResponse Object}
\usage{
LeaderboardListResponse(items = NULL, nextPageToken = NULL)
}
\arguments{
\item{items}{The leaderboards}

\item{nextPageToken}{Token corresponding to the next page of results}
}
\value{
LeaderboardListResponse object
}
\description{
LeaderboardListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for a list of leaderboard objects.
}

