% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{LeaderboardScoreRank}
\alias{LeaderboardScoreRank}
\title{LeaderboardScoreRank Object}
\usage{
LeaderboardScoreRank(formattedNumScores = NULL, formattedRank = NULL,
  numScores = NULL, rank = NULL)
}
\arguments{
\item{formattedNumScores}{The number of scores in the leaderboard as a string}

\item{formattedRank}{The rank in the leaderboard as a string}

\item{numScores}{The number of scores in the leaderboard}

\item{rank}{The rank in the leaderboard}
}
\value{
LeaderboardScoreRank object
}
\description{
LeaderboardScoreRank Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for a score rank in a leaderboard.
}

