% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{LeaderboardScores}
\alias{LeaderboardScores}
\title{LeaderboardScores Object}
\usage{
LeaderboardScores(items = NULL, nextPageToken = NULL, numScores = NULL,
  playerScore = NULL, prevPageToken = NULL)
}
\arguments{
\item{items}{The scores in the leaderboard}

\item{nextPageToken}{The pagination token for the next page of results}

\item{numScores}{The total number of scores in the leaderboard}

\item{playerScore}{The score of the requesting player on the leaderboard}

\item{prevPageToken}{The pagination token for the previous page of results}
}
\value{
LeaderboardScores object
}
\description{
LeaderboardScores Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for a ListScores response.
}
