% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_functions.R
\name{metagame.listCategoriesByPlayer}
\alias{metagame.listCategoriesByPlayer}
\title{List play data aggregated per category for the player corresponding to playerId.}
\usage{
metagame.listCategoriesByPlayer(playerId, collection, consistencyToken = NULL,
  language = NULL, maxResults = NULL, pageToken = NULL)
}
\arguments{
\item{playerId}{A player ID}

\item{collection}{The collection of categories for which data will be returned}

\item{consistencyToken}{The last-seen mutation timestamp}

\item{language}{The preferred language to use for strings returned by this method}

\item{maxResults}{The maximum number of category resources to return in the response, used for paging}

\item{pageToken}{The token returned by the previous request}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/games
\item https://www.googleapis.com/auth/plus.login
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/games, https://www.googleapis.com/auth/plus.login)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/games/services/}{Google Documentation}
}
