% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{MetagameConfig}
\alias{MetagameConfig}
\title{MetagameConfig Object}
\usage{
MetagameConfig(currentVersion = NULL, playerLevels = NULL)
}
\arguments{
\item{currentVersion}{Current version of the metagame configuration data}

\item{playerLevels}{The list of player levels}
}
\value{
MetagameConfig object
}
\description{
MetagameConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for the metagame config resource
}
