% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{NetworkDiagnostics}
\alias{NetworkDiagnostics}
\title{NetworkDiagnostics Object}
\usage{
NetworkDiagnostics(androidNetworkSubtype = NULL, androidNetworkType = NULL,
  iosNetworkType = NULL, networkOperatorCode = NULL,
  networkOperatorName = NULL, registrationLatencyMillis = NULL)
}
\arguments{
\item{androidNetworkSubtype}{The Android network subtype}

\item{androidNetworkType}{The Android network type}

\item{iosNetworkType}{iOS network type as defined in Reachability}

\item{networkOperatorCode}{The MCC+MNC code for the client's network connection}

\item{networkOperatorName}{The name of the carrier of the client's network connection}

\item{registrationLatencyMillis}{The amount of time in milliseconds it took for the client to establish a connection with the XMPP server}
}
\value{
NetworkDiagnostics object
}
\description{
NetworkDiagnostics Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for network diagnostics reported for a client.
}
