% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{ParticipantResult}
\alias{ParticipantResult}
\title{ParticipantResult Object}
\usage{
ParticipantResult(participantId = NULL, placing = NULL, result = NULL)
}
\arguments{
\item{participantId}{The ID of the participant}

\item{placing}{The placement or ranking of the participant in the match results; a number from one to the number of participants in the match}

\item{result}{The result of the participant for this match}
}
\value{
ParticipantResult object
}
\description{
ParticipantResult Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for a result for a match participant.
}
