% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{PeerChannelDiagnostics}
\alias{PeerChannelDiagnostics}
\title{PeerChannelDiagnostics Object}
\usage{
PeerChannelDiagnostics(bytesReceived = NULL, bytesSent = NULL,
  numMessagesLost = NULL, numMessagesReceived = NULL,
  numMessagesSent = NULL, numSendFailures = NULL,
  roundtripLatencyMillis = NULL)
}
\arguments{
\item{bytesReceived}{Number of bytes received}

\item{bytesSent}{Number of bytes sent}

\item{numMessagesLost}{Number of messages lost}

\item{numMessagesReceived}{Number of messages received}

\item{numMessagesSent}{Number of messages sent}

\item{numSendFailures}{Number of send failures}

\item{roundtripLatencyMillis}{Roundtrip latency stats in milliseconds}
}
\value{
PeerChannelDiagnostics object
}
\description{
PeerChannelDiagnostics Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for peer channel diagnostics.
}
