% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{PeerSessionDiagnostics}
\alias{PeerSessionDiagnostics}
\title{PeerSessionDiagnostics Object}
\usage{
PeerSessionDiagnostics(connectedTimestampMillis = NULL,
  participantId = NULL, reliableChannel = NULL, unreliableChannel = NULL)
}
\arguments{
\item{connectedTimestampMillis}{Connected time in milliseconds}

\item{participantId}{The participant ID of the peer}

\item{reliableChannel}{Reliable channel diagnostics}

\item{unreliableChannel}{Unreliable channel diagnostics}
}
\value{
PeerSessionDiagnostics object
}
\description{
PeerSessionDiagnostics Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for peer session diagnostics.
}

