% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{PlayerAchievement}
\alias{PlayerAchievement}
\title{PlayerAchievement Object}
\usage{
PlayerAchievement(achievementState = NULL, currentSteps = NULL,
  experiencePoints = NULL, formattedCurrentStepsString = NULL, id = NULL,
  lastUpdatedTimestamp = NULL)
}
\arguments{
\item{achievementState}{The state of the achievement}

\item{currentSteps}{The current steps for an incremental achievement}

\item{experiencePoints}{Experience points earned for the achievement}

\item{formattedCurrentStepsString}{The current steps for an incremental achievement as a string}

\item{id}{The ID of the achievement}

\item{lastUpdatedTimestamp}{The timestamp of the last modification to this achievement's state}
}
\value{
PlayerAchievement object
}
\description{
PlayerAchievement Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for an achievement object.
}
