% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{PlayerAchievementListResponse}
\alias{PlayerAchievementListResponse}
\title{PlayerAchievementListResponse Object}
\usage{
PlayerAchievementListResponse(items = NULL, nextPageToken = NULL)
}
\arguments{
\item{items}{The achievements}

\item{nextPageToken}{Token corresponding to the next page of results}
}
\value{
PlayerAchievementListResponse object
}
\description{
PlayerAchievementListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for a list of achievement objects.
}

