% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{PlayerEvent}
\alias{PlayerEvent}
\title{PlayerEvent Object}
\usage{
PlayerEvent(definitionId = NULL, formattedNumEvents = NULL,
  numEvents = NULL, playerId = NULL)
}
\arguments{
\item{definitionId}{The ID of the event definition}

\item{formattedNumEvents}{The current number of times this event has occurred, as a string}

\item{numEvents}{The current number of times this event has occurred}

\item{playerId}{The ID of the player}
}
\value{
PlayerEvent object
}
\description{
PlayerEvent Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for an event status resource.
}
