% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{PlayerEventListResponse}
\alias{PlayerEventListResponse}
\title{PlayerEventListResponse Object}
\usage{
PlayerEventListResponse(items = NULL, nextPageToken = NULL)
}
\arguments{
\item{items}{The player events}

\item{nextPageToken}{The pagination token for the next page of results}
}
\value{
PlayerEventListResponse object
}
\description{
PlayerEventListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for a ListByPlayer response.
}
