% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{PlayerLeaderboardScore}
\alias{PlayerLeaderboardScore}
\title{PlayerLeaderboardScore Object}
\usage{
PlayerLeaderboardScore(leaderboard_id = NULL, publicRank = NULL,
  scoreString = NULL, scoreTag = NULL, scoreValue = NULL,
  socialRank = NULL, timeSpan = NULL, writeTimestamp = NULL)
}
\arguments{
\item{leaderboard_id}{The ID of the leaderboard this score is in}

\item{publicRank}{The public rank of the score in this leaderboard}

\item{scoreString}{The formatted value of this score}

\item{scoreTag}{Additional information about the score}

\item{scoreValue}{The numerical value of this score}

\item{socialRank}{The social rank of the score in this leaderboard}

\item{timeSpan}{The time span of this score}

\item{writeTimestamp}{The timestamp at which this score was recorded, in milliseconds since the epoch in UTC}
}
\value{
PlayerLeaderboardScore object
}
\description{
PlayerLeaderboardScore Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for a player leaderboard score object.
}
