% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{PlayerLeaderboardScoreListResponse}
\alias{PlayerLeaderboardScoreListResponse}
\title{PlayerLeaderboardScoreListResponse Object}
\usage{
PlayerLeaderboardScoreListResponse(items = NULL, nextPageToken = NULL,
  player = NULL)
}
\arguments{
\item{items}{The leaderboard scores}

\item{nextPageToken}{The pagination token for the next page of results}

\item{player}{The Player resources for the owner of this score}
}
\value{
PlayerLeaderboardScoreListResponse object
}
\description{
PlayerLeaderboardScoreListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for a list of player leaderboard scores.
}
