% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{PlayerLevel}
\alias{PlayerLevel}
\title{PlayerLevel Object}
\usage{
PlayerLevel(level = NULL, maxExperiencePoints = NULL,
  minExperiencePoints = NULL)
}
\arguments{
\item{level}{The level for the user}

\item{maxExperiencePoints}{The maximum experience points for this level}

\item{minExperiencePoints}{The minimum experience points for this level}
}
\value{
PlayerLevel object
}
\description{
PlayerLevel Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for 1P/3P metadata about a user's level.
}
