% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{PlayerListResponse}
\alias{PlayerListResponse}
\title{PlayerListResponse Object}
\usage{
PlayerListResponse(items = NULL, nextPageToken = NULL)
}
\arguments{
\item{items}{The players}

\item{nextPageToken}{Token corresponding to the next page of results}
}
\value{
PlayerListResponse object
}
\description{
PlayerListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for a third party player list response.
}

