% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{PlayerScore}
\alias{PlayerScore}
\title{PlayerScore Object}
\usage{
PlayerScore(formattedScore = NULL, score = NULL, scoreTag = NULL,
  timeSpan = NULL)
}
\arguments{
\item{formattedScore}{The formatted score for this player score}

\item{score}{The numerical value for this player score}

\item{scoreTag}{Additional information about this score}

\item{timeSpan}{The time span for this player score}
}
\value{
PlayerScore object
}
\description{
PlayerScore Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for a player score.
}

