% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{PlayerScoreListResponse}
\alias{PlayerScoreListResponse}
\title{PlayerScoreListResponse Object}
\usage{
PlayerScoreListResponse(submittedScores = NULL)
}
\arguments{
\item{submittedScores}{The score submissions statuses}
}
\value{
PlayerScoreListResponse object
}
\description{
PlayerScoreListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for a list of score submission statuses.
}

