% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{PlayerScoreResponse}
\alias{PlayerScoreResponse}
\title{PlayerScoreResponse Object}
\usage{
PlayerScoreResponse(beatenScoreTimeSpans = NULL, formattedScore = NULL,
  leaderboardId = NULL, scoreTag = NULL, unbeatenScores = NULL)
}
\arguments{
\item{beatenScoreTimeSpans}{The time spans where the submitted score is better than the existing score for that time span}

\item{formattedScore}{The formatted value of the submitted score}

\item{leaderboardId}{The leaderboard ID that this score was submitted to}

\item{scoreTag}{Additional information about this score}

\item{unbeatenScores}{The scores in time spans that have not been beaten}
}
\value{
PlayerScoreResponse object
}
\description{
PlayerScoreResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for a list of leaderboard entry resources.
}

