% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{PlayerScoreSubmissionList}
\alias{PlayerScoreSubmissionList}
\title{PlayerScoreSubmissionList Object}
\usage{
PlayerScoreSubmissionList(scores = NULL)
}
\arguments{
\item{scores}{The score submissions}
}
\value{
PlayerScoreSubmissionList object
}
\description{
PlayerScoreSubmissionList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for a list of score submission requests
}
\seealso{
Other PlayerScoreSubmissionList functions: \code{\link{scores.submitMultiple}}
}
