% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{ProfileSettings}
\alias{ProfileSettings}
\title{ProfileSettings Object}
\usage{
ProfileSettings(profileVisible = NULL)
}
\arguments{
\item{profileVisible}{The player's current profile visibility}
}
\value{
ProfileSettings object
}
\description{
ProfileSettings Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for profile settings
}
