% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{PushToken}
\alias{PushToken}
\title{PushToken Object}
\usage{
PushToken(clientRevision = NULL, id = NULL, language = NULL)
}
\arguments{
\item{clientRevision}{The revision of the client SDK used by your application, in the same format that's used by revisions}

\item{id}{Unique identifier for this push token}

\item{language}{The preferred language for notifications that are sent using this token}
}
\value{
PushToken object
}
\description{
PushToken Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for a push token resource.
}
\seealso{
Other PushToken functions: \code{\link{pushtokens.update}}
}
