% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{PushTokenId.ios}
\alias{PushTokenId.ios}
\title{PushTokenId.ios Object}
\usage{
PushTokenId.ios(apns_device_token = NULL, apns_environment = NULL)
}
\arguments{
\item{apns_device_token}{Device token supplied by an iOS system call to register for remote notifications}

\item{apns_environment}{Indicates whether this token should be used for the production or sandbox APNS server}
}
\value{
PushTokenId.ios object
}
\description{
PushTokenId.ios Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A push token ID for iOS devices.
}
\seealso{
Other PushTokenId functions: \code{\link{PushTokenId}},
  \code{\link{pushtokens.remove}}
}

