% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{PushTokenId}
\alias{PushTokenId}
\title{PushTokenId Object}
\usage{
PushTokenId(PushTokenId.ios = NULL, ios = NULL)
}
\arguments{
\item{PushTokenId.ios}{The \link{PushTokenId.ios} object or list of objects}

\item{ios}{A push token ID for iOS devices}
}
\value{
PushTokenId object
}
\description{
PushTokenId Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for a push token ID resource.
}
\seealso{
Other PushTokenId functions: \code{\link{PushTokenId.ios}},
  \code{\link{pushtokens.remove}}
}
