% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{Quest}
\alias{Quest}
\title{Quest Object}
\usage{
Quest(acceptedTimestampMillis = NULL, applicationId = NULL,
  bannerUrl = NULL, description = NULL, endTimestampMillis = NULL,
  iconUrl = NULL, id = NULL, isDefaultBannerUrl = NULL,
  isDefaultIconUrl = NULL, lastUpdatedTimestampMillis = NULL,
  milestones = NULL, name = NULL, notifyTimestampMillis = NULL,
  startTimestampMillis = NULL, state = NULL)
}
\arguments{
\item{acceptedTimestampMillis}{The timestamp at which the user accepted the quest in milliseconds since the epoch in UTC}

\item{applicationId}{The ID of the application this quest is part of}

\item{bannerUrl}{The banner image URL for the quest}

\item{description}{The description of the quest}

\item{endTimestampMillis}{The timestamp at which the quest ceases to be active in milliseconds since the epoch in UTC}

\item{iconUrl}{The icon image URL for the quest}

\item{id}{The ID of the quest}

\item{isDefaultBannerUrl}{Indicates whether the banner image being returned is a default image, or is game-provided}

\item{isDefaultIconUrl}{Indicates whether the icon image being returned is a default image, or is game-provided}

\item{lastUpdatedTimestampMillis}{The timestamp at which the quest was last updated by the user in milliseconds since the epoch in UTC}

\item{milestones}{The quest milestones}

\item{name}{The name of the quest}

\item{notifyTimestampMillis}{The timestamp at which the user should be notified that the quest will end soon in milliseconds since the epoch in UTC}

\item{startTimestampMillis}{The timestamp at which the quest becomes active in milliseconds since the epoch in UTC}

\item{state}{The state of the quest}
}
\value{
Quest object
}
\description{
Quest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for a Quest resource.
}
