% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{QuestContribution}
\alias{QuestContribution}
\title{QuestContribution Object}
\usage{
QuestContribution(formattedValue = NULL, value = NULL)
}
\arguments{
\item{formattedValue}{The formatted value of the contribution as a string}

\item{value}{The value of the contribution}
}
\value{
QuestContribution object
}
\description{
QuestContribution Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for a Quest Criterion Contribution resource.
}
