% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{QuestCriterion}
\alias{QuestCriterion}
\title{QuestCriterion Object}
\usage{
QuestCriterion(completionContribution = NULL, currentContribution = NULL,
  eventId = NULL, initialPlayerProgress = NULL)
}
\arguments{
\item{completionContribution}{The total number of times the associated event must be incremented for the player to complete this quest}

\item{currentContribution}{The number of increments the player has made toward the completion count event increments required to complete the quest}

\item{eventId}{The ID of the event the criterion corresponds to}

\item{initialPlayerProgress}{The value of the event associated with this quest at the time that the quest was accepted}
}
\value{
QuestCriterion object
}
\description{
QuestCriterion Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for a Quest Criterion resource.
}

