% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{QuestListResponse}
\alias{QuestListResponse}
\title{QuestListResponse Object}
\usage{
QuestListResponse(items = NULL, nextPageToken = NULL)
}
\arguments{
\item{items}{The quests}

\item{nextPageToken}{Token corresponding to the next page of results}
}
\value{
QuestListResponse object
}
\description{
QuestListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for a list of quest objects.
}
