% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{QuestMilestone}
\alias{QuestMilestone}
\title{QuestMilestone Object}
\usage{
QuestMilestone(completionRewardData = NULL, criteria = NULL, id = NULL,
  state = NULL)
}
\arguments{
\item{completionRewardData}{The completion reward data of the milestone, represented as a Base64-encoded string}

\item{criteria}{The criteria of the milestone}

\item{id}{The milestone ID}

\item{state}{The current state of the milestone}
}
\value{
QuestMilestone object
}
\description{
QuestMilestone Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for a Quest Milestone resource.
}
