% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_functions.R
\name{questMilestones.claim}
\alias{questMilestones.claim}
\title{Report that a reward for the milestone corresponding to milestoneId for the quest corresponding to questId has been claimed by the currently authorized user.}
\usage{
questMilestones.claim(questId, milestoneId, requestId,
  consistencyToken = NULL)
}
\arguments{
\item{questId}{The ID of the quest}

\item{milestoneId}{The ID of the milestone}

\item{requestId}{A numeric ID to ensure that the request is handled correctly across retries}

\item{consistencyToken}{The last-seen mutation timestamp}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/games
\item https://www.googleapis.com/auth/plus.login
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/games, https://www.googleapis.com/auth/plus.login)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/games/services/}{Google Documentation}
}

