% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_functions.R
\name{quests.accept}
\alias{quests.accept}
\title{Indicates that the currently authorized user will participate in the quest.}
\usage{
quests.accept(questId, consistencyToken = NULL, language = NULL)
}
\arguments{
\item{questId}{The ID of the quest}

\item{consistencyToken}{The last-seen mutation timestamp}

\item{language}{The preferred language to use for strings returned by this method}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/games
\item https://www.googleapis.com/auth/plus.login
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/games, https://www.googleapis.com/auth/plus.login)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/games/services/}{Google Documentation}
}
