% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{RevisionCheckResponse}
\alias{RevisionCheckResponse}
\title{RevisionCheckResponse Object}
\usage{
RevisionCheckResponse(apiVersion = NULL, revisionStatus = NULL)
}
\arguments{
\item{apiVersion}{The version of the API this client revision should use when calling API methods}

\item{revisionStatus}{The result of the revision check}
}
\value{
RevisionCheckResponse object
}
\description{
RevisionCheckResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for the result of checking a revision.
}

