% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{Room}
\alias{Room}
\title{Room Object}
\usage{
Room(applicationId = NULL, autoMatchingCriteria = NULL,
  autoMatchingStatus = NULL, creationDetails = NULL, description = NULL,
  inviterId = NULL, lastUpdateDetails = NULL, participants = NULL,
  roomId = NULL, roomStatusVersion = NULL, status = NULL,
  variant = NULL)
}
\arguments{
\item{applicationId}{The ID of the application being played}

\item{autoMatchingCriteria}{Criteria for auto-matching players into this room}

\item{autoMatchingStatus}{Auto-matching status for this room}

\item{creationDetails}{Details about the room creation}

\item{description}{This short description is generated by our servers and worded relative to the player requesting the room}

\item{inviterId}{The ID of the participant that invited the user to the room}

\item{lastUpdateDetails}{Details about the last update to the room}

\item{participants}{The participants involved in the room, along with their statuses}

\item{roomId}{Globally unique ID for a room}

\item{roomStatusVersion}{The version of the room status: an increasing counter, used by the client to ignore out-of-order updates to room status}

\item{status}{The status of the room}

\item{variant}{The variant / mode of the application being played; can be any integer value, or left blank}
}
\value{
Room object
}
\description{
Room Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for a room resource object.
}
