% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{RoomAutoMatchingCriteria}
\alias{RoomAutoMatchingCriteria}
\title{RoomAutoMatchingCriteria Object}
\usage{
RoomAutoMatchingCriteria(exclusiveBitmask = NULL,
  maxAutoMatchingPlayers = NULL, minAutoMatchingPlayers = NULL)
}
\arguments{
\item{exclusiveBitmask}{A bitmask indicating when auto-matches are valid}

\item{maxAutoMatchingPlayers}{The maximum number of players that should be added to the room by auto-matching}

\item{minAutoMatchingPlayers}{The minimum number of players that should be added to the room by auto-matching}
}
\value{
RoomAutoMatchingCriteria object
}
\description{
RoomAutoMatchingCriteria Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for a room auto-match criteria object.
}

