% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{RoomAutoMatchStatus}
\alias{RoomAutoMatchStatus}
\title{RoomAutoMatchStatus Object}
\usage{
RoomAutoMatchStatus(waitEstimateSeconds = NULL)
}
\arguments{
\item{waitEstimateSeconds}{An estimate for the amount of time (in seconds) that auto-matching is expected to take to complete}
}
\value{
RoomAutoMatchStatus object
}
\description{
RoomAutoMatchStatus Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for status of room automatching that is in progress.
}

