% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{RoomClientAddress}
\alias{RoomClientAddress}
\title{RoomClientAddress Object}
\usage{
RoomClientAddress(xmppAddress = NULL)
}
\arguments{
\item{xmppAddress}{The XMPP address of the client on the Google Games XMPP network}
}
\value{
RoomClientAddress object
}
\description{
RoomClientAddress Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for the client address when setting up a room.
}

