% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{RoomCreateRequest}
\alias{RoomCreateRequest}
\title{RoomCreateRequest Object}
\usage{
RoomCreateRequest(autoMatchingCriteria = NULL, capabilities = NULL,
  clientAddress = NULL, invitedPlayerIds = NULL,
  networkDiagnostics = NULL, requestId = NULL, variant = NULL)
}
\arguments{
\item{autoMatchingCriteria}{Criteria for auto-matching players into this room}

\item{capabilities}{The capabilities that this client supports for realtime communication}

\item{clientAddress}{Client address for the player creating the room}

\item{invitedPlayerIds}{The player IDs to invite to the room}

\item{networkDiagnostics}{Network diagnostics for the client creating the room}

\item{requestId}{A randomly generated numeric ID}

\item{variant}{The variant / mode of the application to be played}
}
\value{
RoomCreateRequest object
}
\description{
RoomCreateRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for a room creation request.
}
\seealso{
Other RoomCreateRequest functions: \code{\link{rooms.create}}
}
