% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{RoomJoinRequest}
\alias{RoomJoinRequest}
\title{RoomJoinRequest Object}
\usage{
RoomJoinRequest(capabilities = NULL, clientAddress = NULL,
  networkDiagnostics = NULL)
}
\arguments{
\item{capabilities}{The capabilities that this client supports for realtime communication}

\item{clientAddress}{Client address for the player joining the room}

\item{networkDiagnostics}{Network diagnostics for the client joining the room}
}
\value{
RoomJoinRequest object
}
\description{
RoomJoinRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for a join room request.
}
\seealso{
Other RoomJoinRequest functions: \code{\link{rooms.join}}
}

