% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{RoomLeaveDiagnostics}
\alias{RoomLeaveDiagnostics}
\title{RoomLeaveDiagnostics Object}
\usage{
RoomLeaveDiagnostics(androidNetworkSubtype = NULL,
  androidNetworkType = NULL, iosNetworkType = NULL,
  networkOperatorCode = NULL, networkOperatorName = NULL,
  peerSession = NULL, socketsUsed = NULL)
}
\arguments{
\item{androidNetworkSubtype}{Android network subtype}

\item{androidNetworkType}{Android network type}

\item{iosNetworkType}{iOS network type as defined in Reachability}

\item{networkOperatorCode}{The MCC+MNC code for the client's network connection}

\item{networkOperatorName}{The name of the carrier of the client's network connection}

\item{peerSession}{Diagnostics about all peer sessions}

\item{socketsUsed}{Whether or not sockets were used}
}
\value{
RoomLeaveDiagnostics object
}
\description{
RoomLeaveDiagnostics Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for room leave diagnostics.
}

