% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{RoomLeaveRequest}
\alias{RoomLeaveRequest}
\title{RoomLeaveRequest Object}
\usage{
RoomLeaveRequest(leaveDiagnostics = NULL, reason = NULL)
}
\arguments{
\item{leaveDiagnostics}{Diagnostics for a player leaving the room}

\item{reason}{Reason for leaving the match}
}
\value{
RoomLeaveRequest object
}
\description{
RoomLeaveRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for a leave room request.
}
\seealso{
Other RoomLeaveRequest functions: \code{\link{rooms.leave}}
}

