% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{RoomList}
\alias{RoomList}
\title{RoomList Object}
\usage{
RoomList(items = NULL, nextPageToken = NULL)
}
\arguments{
\item{items}{The rooms}

\item{nextPageToken}{The pagination token for the next page of results}
}
\value{
RoomList object
}
\description{
RoomList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for a list of rooms.
}

