% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{RoomModification}
\alias{RoomModification}
\title{RoomModification Object}
\usage{
RoomModification(modifiedTimestampMillis = NULL, participantId = NULL)
}
\arguments{
\item{modifiedTimestampMillis}{The timestamp at which they modified the room, in milliseconds since the epoch in UTC}

\item{participantId}{The ID of the participant that modified the room}
}
\value{
RoomModification object
}
\description{
RoomModification Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for room modification metadata.
}
