% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{RoomP2PStatus}
\alias{RoomP2PStatus}
\title{RoomP2PStatus Object}
\usage{
RoomP2PStatus(connectionSetupLatencyMillis = NULL, error = NULL,
  error_reason = NULL, participantId = NULL, status = NULL,
  unreliableRoundtripLatencyMillis = NULL)
}
\arguments{
\item{connectionSetupLatencyMillis}{The amount of time in milliseconds it took to establish connections with this peer}

\item{error}{The error code in event of a failure}

\item{error_reason}{More detailed diagnostic message returned in event of a failure}

\item{participantId}{The ID of the participant}

\item{status}{The status of the peer in the room}

\item{unreliableRoundtripLatencyMillis}{The amount of time in milliseconds it took to send packets back and forth on the unreliable channel with this peer}
}
\value{
RoomP2PStatus object
}
\description{
RoomP2PStatus Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for an update on the status of a peer in a room.
}
