% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{RoomP2PStatuses}
\alias{RoomP2PStatuses}
\title{RoomP2PStatuses Object}
\usage{
RoomP2PStatuses(updates = NULL)
}
\arguments{
\item{updates}{The updates for the peers}
}
\value{
RoomP2PStatuses object
}
\description{
RoomP2PStatuses Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for an update on the status of peers in a room.
}
\seealso{
Other RoomP2PStatuses functions: \code{\link{rooms.reportStatus}}
}
