% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{RoomParticipant}
\alias{RoomParticipant}
\title{RoomParticipant Object}
\usage{
RoomParticipant(autoMatched = NULL, autoMatchedPlayer = NULL,
  capabilities = NULL, clientAddress = NULL, connected = NULL,
  id = NULL, leaveReason = NULL, player = NULL, status = NULL)
}
\arguments{
\item{autoMatched}{True if this participant was auto-matched with the requesting player}

\item{autoMatchedPlayer}{Information about a player that has been anonymously auto-matched against the requesting player}

\item{capabilities}{The capabilities which can be used when communicating with this participant}

\item{clientAddress}{Client address for the participant}

\item{connected}{True if this participant is in the fully connected set of peers in the room}

\item{id}{An identifier for the participant in the scope of the room}

\item{leaveReason}{The reason the participant left the room; populated if the participant status is PARTICIPANT_LEFT}

\item{player}{Information about the player}

\item{status}{The status of the participant with respect to the room}
}
\value{
RoomParticipant object
}
\description{
RoomParticipant Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for a participant in a room.
}
