% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_functions.R
\name{rooms.dismiss}
\alias{rooms.dismiss}
\title{Dismiss an invitation to join a room. For internal use by the Games SDK only. Calling this method directly is unsupported.}
\usage{
rooms.dismiss(roomId, consistencyToken = NULL)
}
\arguments{
\item{roomId}{The ID of the room}

\item{consistencyToken}{The last-seen mutation timestamp}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/games
\item https://www.googleapis.com/auth/plus.login
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/games, https://www.googleapis.com/auth/plus.login)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/games/services/}{Google Documentation}
}

