% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_functions.R
\name{rooms.leave}
\alias{rooms.leave}
\title{Leave a room. For internal use by the Games SDK only. Calling this method directly is unsupported.}
\usage{
rooms.leave(RoomLeaveRequest, roomId, consistencyToken = NULL,
  language = NULL)
}
\arguments{
\item{RoomLeaveRequest}{The \link{RoomLeaveRequest} object to pass to this method}

\item{roomId}{The ID of the room}

\item{consistencyToken}{The last-seen mutation timestamp}

\item{language}{The preferred language to use for strings returned by this method}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/games
\item https://www.googleapis.com/auth/plus.login
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/games, https://www.googleapis.com/auth/plus.login)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/games/services/}{Google Documentation}

Other RoomLeaveRequest functions: \code{\link{RoomLeaveRequest}}
}

