% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{RoomStatus}
\alias{RoomStatus}
\title{RoomStatus Object}
\usage{
RoomStatus(autoMatchingStatus = NULL, participants = NULL, roomId = NULL,
  status = NULL, statusVersion = NULL)
}
\arguments{
\item{autoMatchingStatus}{Auto-matching status for this room}

\item{participants}{The participants involved in the room, along with their statuses}

\item{roomId}{Globally unique ID for a room}

\item{status}{The status of the room}

\item{statusVersion}{The version of the status for the room: an increasing counter, used by the client to ignore out-of-order updates to room status}
}
\value{
RoomStatus object
}
\description{
RoomStatus Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for the status of a room that the player has joined.
}
