% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_functions.R
\name{scores.get}
\alias{scores.get}
\title{Get high scores, and optionally ranks, in leaderboards for the currently authenticated player. For a specific time span, leaderboardId can be set to ALL to retrieve data for all leaderboards in a given time span.NOTE: You cannot ask for 'ALL' leaderboards and 'ALL' timeSpans in the same request; only one parameter may be set to 'ALL'.}
\usage{
scores.get(playerId, leaderboardId, timeSpan, consistencyToken = NULL,
  includeRankType = NULL, language = NULL, maxResults = NULL,
  pageToken = NULL)
}
\arguments{
\item{playerId}{A player ID}

\item{leaderboardId}{The ID of the leaderboard}

\item{timeSpan}{The time span for the scores and ranks you're requesting}

\item{consistencyToken}{The last-seen mutation timestamp}

\item{includeRankType}{The types of ranks to return}

\item{language}{The preferred language to use for strings returned by this method}

\item{maxResults}{The maximum number of leaderboard scores to return in the response}

\item{pageToken}{The token returned by the previous request}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/games
\item https://www.googleapis.com/auth/plus.login
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/games, https://www.googleapis.com/auth/plus.login)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/games/services/}{Google Documentation}
}
