% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_functions.R
\name{scores.list}
\alias{scores.list}
\title{Lists the scores in a leaderboard, starting from the top.}
\usage{
scores.list(leaderboardId, collection, timeSpan, consistencyToken = NULL,
  language = NULL, maxResults = NULL, pageToken = NULL)
}
\arguments{
\item{leaderboardId}{The ID of the leaderboard}

\item{collection}{The collection of scores you're requesting}

\item{timeSpan}{The time span for the scores and ranks you're requesting}

\item{consistencyToken}{The last-seen mutation timestamp}

\item{language}{The preferred language to use for strings returned by this method}

\item{maxResults}{The maximum number of leaderboard scores to return in the response}

\item{pageToken}{The token returned by the previous request}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/games
\item https://www.googleapis.com/auth/plus.login
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/games, https://www.googleapis.com/auth/plus.login)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/games/services/}{Google Documentation}
}

