% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_functions.R
\name{scores.submit}
\alias{scores.submit}
\title{Submits a score to the specified leaderboard.}
\usage{
scores.submit(leaderboardId, score, consistencyToken = NULL,
  language = NULL, scoreTag = NULL)
}
\arguments{
\item{leaderboardId}{The ID of the leaderboard}

\item{score}{The score you're submitting}

\item{consistencyToken}{The last-seen mutation timestamp}

\item{language}{The preferred language to use for strings returned by this method}

\item{scoreTag}{Additional information about the score you're submitting}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/games
\item https://www.googleapis.com/auth/plus.login
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/games, https://www.googleapis.com/auth/plus.login)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/games/services/}{Google Documentation}
}
