% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{ScoreSubmission}
\alias{ScoreSubmission}
\title{ScoreSubmission Object}
\usage{
ScoreSubmission(leaderboardId = NULL, score = NULL, scoreTag = NULL,
  signature = NULL)
}
\arguments{
\item{leaderboardId}{The leaderboard this score is being submitted to}

\item{score}{The new score being submitted}

\item{scoreTag}{Additional information about this score}

\item{signature}{Signature Values will contain URI-safe characters as defined by section 2}
}
\value{
ScoreSubmission object
}
\description{
ScoreSubmission Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for a request to submit a score to leaderboards.
}

