% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{Snapshot}
\alias{Snapshot}
\title{Snapshot Object}
\usage{
Snapshot(coverImage = NULL, description = NULL, driveId = NULL,
  durationMillis = NULL, id = NULL, lastModifiedMillis = NULL,
  progressValue = NULL, title = NULL, type = NULL, uniqueName = NULL)
}
\arguments{
\item{coverImage}{The cover image of this snapshot}

\item{description}{The description of this snapshot}

\item{driveId}{The ID of the file underlying this snapshot in the Drive API}

\item{durationMillis}{The duration associated with this snapshot, in millis}

\item{id}{The ID of the snapshot}

\item{lastModifiedMillis}{The timestamp (in millis since Unix epoch) of the last modification to this snapshot}

\item{progressValue}{The progress value (64-bit integer set by developer) associated with this snapshot}

\item{title}{The title of this snapshot}

\item{type}{The type of this snapshot}

\item{uniqueName}{The unique name provided when the snapshot was created}
}
\value{
Snapshot object
}
\description{
Snapshot Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for an snapshot object.
}
