% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{SnapshotImage}
\alias{SnapshotImage}
\title{SnapshotImage Object}
\usage{
SnapshotImage(height = NULL, mime_type = NULL, url = NULL, width = NULL)
}
\arguments{
\item{height}{The height of the image}

\item{mime_type}{The MIME type of the image}

\item{url}{The URL of the image}

\item{width}{The width of the image}
}
\value{
SnapshotImage object
}
\description{
SnapshotImage Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for an image of a snapshot.
}
