% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{SnapshotListResponse}
\alias{SnapshotListResponse}
\title{SnapshotListResponse Object}
\usage{
SnapshotListResponse(items = NULL, nextPageToken = NULL)
}
\arguments{
\item{items}{The snapshots}

\item{nextPageToken}{Token corresponding to the next page of results}
}
\value{
SnapshotListResponse object
}
\description{
SnapshotListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for a list of snapshot objects.
}

