% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{TurnBasedAutoMatchingCriteria}
\alias{TurnBasedAutoMatchingCriteria}
\title{TurnBasedAutoMatchingCriteria Object}
\usage{
TurnBasedAutoMatchingCriteria(exclusiveBitmask = NULL,
  maxAutoMatchingPlayers = NULL, minAutoMatchingPlayers = NULL)
}
\arguments{
\item{exclusiveBitmask}{A bitmask indicating when auto-matches are valid}

\item{maxAutoMatchingPlayers}{The maximum number of players that should be added to the match by auto-matching}

\item{minAutoMatchingPlayers}{The minimum number of players that should be added to the match by auto-matching}
}
\value{
TurnBasedAutoMatchingCriteria object
}
\description{
TurnBasedAutoMatchingCriteria Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for an turn-based auto-match criteria object.
}

