% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{TurnBasedMatch}
\alias{TurnBasedMatch}
\title{TurnBasedMatch Object}
\usage{
TurnBasedMatch(applicationId = NULL, autoMatchingCriteria = NULL,
  creationDetails = NULL, data = NULL, description = NULL,
  inviterId = NULL, lastUpdateDetails = NULL, matchId = NULL,
  matchNumber = NULL, matchVersion = NULL, participants = NULL,
  pendingParticipantId = NULL, previousMatchData = NULL, rematchId = NULL,
  results = NULL, status = NULL, userMatchStatus = NULL, variant = NULL,
  withParticipantId = NULL)
}
\arguments{
\item{applicationId}{The ID of the application being played}

\item{autoMatchingCriteria}{Criteria for auto-matching players into this match}

\item{creationDetails}{Details about the match creation}

\item{data}{The data / game state for this match}

\item{description}{This short description is generated by our servers based on turn state and is localized and worded relative to the player requesting the match}

\item{inviterId}{The ID of the participant that invited the user to the match}

\item{lastUpdateDetails}{Details about the last update to the match}

\item{matchId}{Globally unique ID for a turn-based match}

\item{matchNumber}{The number of the match in a chain of rematches}

\item{matchVersion}{The version of this match: an increasing counter, used to avoid out-of-date updates to the match}

\item{participants}{The participants involved in the match, along with their statuses}

\item{pendingParticipantId}{The ID of the participant that is taking a turn}

\item{previousMatchData}{The data / game state for the previous match; set for the first turn of rematches only}

\item{rematchId}{The ID of a rematch of this match}

\item{results}{The results reported for this match}

\item{status}{The status of the match}

\item{userMatchStatus}{The status of the current user in the match}

\item{variant}{The variant / mode of the application being played; can be any integer value, or left blank}

\item{withParticipantId}{The ID of another participant in the match that can be used when describing the participants the user is playing with}
}
\value{
TurnBasedMatch object
}
\description{
TurnBasedMatch Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for a turn-based match resource object.
}

