% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{TurnBasedMatchCreateRequest}
\alias{TurnBasedMatchCreateRequest}
\title{TurnBasedMatchCreateRequest Object}
\usage{
TurnBasedMatchCreateRequest(autoMatchingCriteria = NULL,
  invitedPlayerIds = NULL, requestId = NULL, variant = NULL)
}
\arguments{
\item{autoMatchingCriteria}{Criteria for auto-matching players into this match}

\item{invitedPlayerIds}{The player ids to invite to the match}

\item{requestId}{A randomly generated numeric ID}

\item{variant}{The variant / mode of the application to be played}
}
\value{
TurnBasedMatchCreateRequest object
}
\description{
TurnBasedMatchCreateRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for a turn-based match creation request.
}
\seealso{
Other TurnBasedMatchCreateRequest functions: \code{\link{turnBasedMatches.create}}
}
