% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{TurnBasedMatchData}
\alias{TurnBasedMatchData}
\title{TurnBasedMatchData Object}
\usage{
TurnBasedMatchData(data = NULL, dataAvailable = NULL)
}
\arguments{
\item{data}{The byte representation of the data (limited to 128 kB), as a Base64-encoded string with the URL_SAFE encoding option}

\item{dataAvailable}{True if this match has data available but it wasn't returned in a list response; fetching the match individually will retrieve this data}
}
\value{
TurnBasedMatchData object
}
\description{
TurnBasedMatchData Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for a turn-based match data object.
}

