% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{TurnBasedMatchDataRequest}
\alias{TurnBasedMatchDataRequest}
\title{TurnBasedMatchDataRequest Object}
\usage{
TurnBasedMatchDataRequest(data = NULL)
}
\arguments{
\item{data}{The byte representation of the data (limited to 128 kB), as a Base64-encoded string with the URL_SAFE encoding option}
}
\value{
TurnBasedMatchDataRequest object
}
\description{
TurnBasedMatchDataRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for sending a turn-based match data object.
}

