% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_functions.R
\name{turnBasedMatches.leaveTurn}
\alias{turnBasedMatches.leaveTurn}
\title{Leave a turn-based match during the current player's turn, without canceling the match.}
\usage{
turnBasedMatches.leaveTurn(matchId, matchVersion, consistencyToken = NULL,
  language = NULL, pendingParticipantId = NULL)
}
\arguments{
\item{matchId}{The ID of the match}

\item{matchVersion}{The version of the match being updated}

\item{consistencyToken}{The last-seen mutation timestamp}

\item{language}{The preferred language to use for strings returned by this method}

\item{pendingParticipantId}{The ID of another participant who should take their turn next}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/games
\item https://www.googleapis.com/auth/plus.login
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/games, https://www.googleapis.com/auth/plus.login)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/games/services/}{Google Documentation}
}

