% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_functions.R
\name{turnBasedMatches.sync}
\alias{turnBasedMatches.sync}
\title{Returns turn-based matches the player is or was involved in that changed since the last sync call, with the least recent changes coming first. Matches that should be removed from the local cache will have a status of MATCH_DELETED.}
\usage{
turnBasedMatches.sync(consistencyToken = NULL, includeMatchData = NULL,
  language = NULL, maxCompletedMatches = NULL, maxResults = NULL,
  pageToken = NULL)
}
\arguments{
\item{consistencyToken}{The last-seen mutation timestamp}

\item{includeMatchData}{True if match data should be returned in the response}

\item{language}{The preferred language to use for strings returned by this method}

\item{maxCompletedMatches}{The maximum number of completed or canceled matches to return in the response}

\item{maxResults}{The maximum number of matches to return in the response, used for paging}

\item{pageToken}{The token returned by the previous request}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/games
\item https://www.googleapis.com/auth/plus.login
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/games, https://www.googleapis.com/auth/plus.login)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/games/services/}{Google Documentation}
}
