% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{TurnBasedMatchList}
\alias{TurnBasedMatchList}
\title{TurnBasedMatchList Object}
\usage{
TurnBasedMatchList(items = NULL, nextPageToken = NULL)
}
\arguments{
\item{items}{The matches}

\item{nextPageToken}{The pagination token for the next page of results}
}
\value{
TurnBasedMatchList object
}
\description{
TurnBasedMatchList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for a list of turn-based matches.
}

