% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{TurnBasedMatchModification}
\alias{TurnBasedMatchModification}
\title{TurnBasedMatchModification Object}
\usage{
TurnBasedMatchModification(modifiedTimestampMillis = NULL,
  participantId = NULL)
}
\arguments{
\item{modifiedTimestampMillis}{The timestamp at which they modified the match, in milliseconds since the epoch in UTC}

\item{participantId}{The ID of the participant that modified the match}
}
\value{
TurnBasedMatchModification object
}
\description{
TurnBasedMatchModification Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for turn-based match modification metadata.
}
