% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{TurnBasedMatchParticipant}
\alias{TurnBasedMatchParticipant}
\title{TurnBasedMatchParticipant Object}
\usage{
TurnBasedMatchParticipant(autoMatched = NULL, autoMatchedPlayer = NULL,
  id = NULL, player = NULL, status = NULL)
}
\arguments{
\item{autoMatched}{True if this participant was auto-matched with the requesting player}

\item{autoMatchedPlayer}{Information about a player that has been anonymously auto-matched against the requesting player}

\item{id}{An identifier for the participant in the scope of the match}

\item{player}{Information about the player}

\item{status}{The status of the participant with respect to the match}
}
\value{
TurnBasedMatchParticipant object
}
\description{
TurnBasedMatchParticipant Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for a participant in a turn-based match.
}

