% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{TurnBasedMatchRematch}
\alias{TurnBasedMatchRematch}
\title{TurnBasedMatchRematch Object}
\usage{
TurnBasedMatchRematch(previousMatch = NULL, rematch = NULL)
}
\arguments{
\item{previousMatch}{The old match that the rematch was created from; will be updated such that the rematchId field will point at the new match}

\item{rematch}{The newly created match; a rematch of the old match with the same participants}
}
\value{
TurnBasedMatchRematch object
}
\description{
TurnBasedMatchRematch Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for a rematch response.
}
