% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{TurnBasedMatchResults}
\alias{TurnBasedMatchResults}
\title{TurnBasedMatchResults Object}
\usage{
TurnBasedMatchResults(data = NULL, matchVersion = NULL, results = NULL)
}
\arguments{
\item{data}{The final match data}

\item{matchVersion}{The version of the match being updated}

\item{results}{The match results for the participants in the match}
}
\value{
TurnBasedMatchResults object
}
\description{
TurnBasedMatchResults Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for a turn-based match results object.
}
\seealso{
Other TurnBasedMatchResults functions: \code{\link{turnBasedMatches.finish}}
}

