% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{TurnBasedMatchSync}
\alias{TurnBasedMatchSync}
\title{TurnBasedMatchSync Object}
\usage{
TurnBasedMatchSync(items = NULL, moreAvailable = NULL,
  nextPageToken = NULL)
}
\arguments{
\item{items}{The matches}

\item{moreAvailable}{True if there were more matches available to fetch at the time the response was generated (which were not returned due to page size limits}

\item{nextPageToken}{The pagination token for the next page of results}
}
\value{
TurnBasedMatchSync object
}
\description{
TurnBasedMatchSync Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for a list of turn-based matches returned from a sync.
}

