% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games_objects.R
\name{TurnBasedMatchTurn}
\alias{TurnBasedMatchTurn}
\title{TurnBasedMatchTurn Object}
\usage{
TurnBasedMatchTurn(data = NULL, matchVersion = NULL,
  pendingParticipantId = NULL, results = NULL)
}
\arguments{
\item{data}{The shared game state data after the turn is over}

\item{matchVersion}{The version of this match: an increasing counter, used to avoid out-of-date updates to the match}

\item{pendingParticipantId}{The ID of the participant who should take their turn next}

\item{results}{The match results for the participants in the match}
}
\value{
TurnBasedMatchTurn object
}
\description{
TurnBasedMatchTurn Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This is a JSON template for the object representing a turn.
}
\seealso{
Other TurnBasedMatchTurn functions: \code{\link{turnBasedMatches.takeTurn}}
}
